+++ werkzeug/datastructures.py
@@ -852,6 +852,15 @@ class Headers(object):
     to the constructor for internal usage.
 
     :param defaults: The list of default values for the :class:`Headers`.
+
+    .. versionchanged:: 0.9
+       This data structure now stores unicode values similar to how the
+       multi dicts do it.  The main difference is that bytes can be set as
+       well which will automatically be latin1 decoded.
+
+    .. versionmodified:: 0.9
+       The :meth:`linked` function was removed without replacement as it
+       was an API that does not support the changes to the encoding model.
     """
 
     def __init__(self, defaults=None):