+++ docs/changes.rst
@@ -22,6 +22,12 @@ API Changes
         make better APIs to deal with request and response entity
         bodies.  So far there is no deprecation warning but users are
         strongly encouraged to update.
+    -   The :class:`Headers` and :class:`EnvironHeaders` datastructures
+        are now designed to operate on unicode data.  This is a backwards
+        incomaptible change and was necessary for the Python 3 support.
+    -   The :class:`Headers` object no longer supports in-place operations
+        through the old ``linked`` method.  This has been removed without
+        replacement due to changes on the encoding model.
 
 `0.6.2`
     -   renamed the attribute `implicit_seqence_conversion` attribute of